split converter further in graph_builder
